+++ packages/core/.size-limit.cjs
@@ -12,8 +12,12 @@ function modifyWebpackConfig(config) {
       resolve: {
         fullySpecified: false,
       },
-    }
+    },
   );
+
+  config.externalsPresets = {
+    node: true,
+  };
 
   return config;
 };
@@ -21,7 +25,7 @@ function modifyWebpackConfig(config) {
 module.exports = [
   {
     limit: '30 KB',
-    path: './dist/client/index.js',
+    path: './dist/esm/client/index.js',
     import: '{ onLoad, reviveClientApp }',
     running: false,
     modifyWebpackConfig,
@@ -29,7 +33,7 @@ module.exports = [
   },
   {
     limit: '30 KB',
-    path: './dist/server/index.js',
+    path: './dist/esm/server/index.js',
     import: '*',
     running: false,
     modifyWebpackConfig,
@@ -37,7 +41,7 @@ module.exports = [
   // Test tree-shaking
   {
     limit: '70 B',
-    path: './dist/client/index.js',
+    path: './dist/esm/client/index.js',
     import: '{ Cache }',
     running: false,
     modifyWebpackConfig,