+++ packages/server/setupJest.js
@@ -1,0 +1,8 @@
+import {
+  objectKeepUnmock,
+  setGlobalKeepUnmock,
+  setGlobalMockMethod,
+} from 'to-mock';
+
+setGlobalMockMethod(jest.fn);
+setGlobalKeepUnmock(objectKeepUnmock);