+++ packages/create-ima-app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 18.0.0-rc.11
+
+### Patch Changes
+
+- af6492c9: Add SPA template to `create-ima-app` and update `DocumentView`. Add support for new inject aliases `#{$Scripts}`, `#{$RevivalSettings}`, `#{$Runner}` and `#{$RevivalCache}` to template.
+- 4046f8b1: Updated template dependencies
+
 ## 18.0.0-rc.10
 
 ### Patch Changes